More warnings/minor fixes
git-svn-id: svn://svn.code.sf.net/p/magicseteditor/code/trunk@836 0fc631ac-6414-0410-93d0-97cfa31319b6

+++ src/data/font.cpp
@@ -88,7 +88,7 @@ wxFont Font::toWxFont(double scale) const {
 		}
 	} else if (name().empty()) {
 		font = *wxNORMAL_FONT;
-		font.SetPointSize(size > 1 ? size_i : scale * font.GetPointSize());
+		font.SetPointSize(size > 1 ? size_i : int(scale * font.GetPointSize()));
 		return font;
 	} else if (flags & FONT_ITALIC && !italic_name().empty()) {
 		font = wxFont(size_i, wxFONTFAMILY_DEFAULT, wxFONTSTYLE_NORMAL, weight_i, underline(), italic_name());
